+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed composer draft format detection - [#420](https://github.com/hkdb/aerion/issues/420)
 - Block save all attachments if flatpak - [#384](https://github.com/hkdb/aerion/issues/384)
 - Calendar color render adjustments to alleviate [#406](https://github.com/hkdb/aerion/issues/406)
+- Fixed Microsoft calendar attachments handling - [#370](https://github.com/hkdb/aerion/issues/370)
 
 
 **v0.3.3 - 08-06-2026**